Split login auth effects
