+++ tests/login-oauth-cookie-fallback.test.ts
@@ -4,22 +4,23 @@ import test from 'node:test';
 
 const loginPageSource = readFileSync('frontend/app/(core)/login/page.tsx', 'utf8');
 const loginControllerSource = readFileSync('frontend/app/(core)/login/_hooks/useLoginPageController.ts', 'utf8');
+const loginOAuthCodeExchangeSource = readFileSync('frontend/app/(core)/login/_hooks/useLoginOAuthCodeExchange.ts', 'utf8');
 const fallbackPath = 'frontend/app/(core)/login/_lib/oauth-cookie-fallback.ts';
 
 test('login keeps OAuth cookie fallback out of page orchestration', () => {
   assert.ok(existsSync(fallbackPath), 'OAuth cookie fallback helper should be route-local');
   assert.match(
-    loginControllerSource,
+    loginOAuthCodeExchangeSource,
     /startOAuthCookieRedirectFallback\(/,
-    'login controller should delegate stalled OAuth cookie detection to a route-local helper'
+    'login OAuth code exchange hook should delegate stalled OAuth cookie detection to a route-local helper'
   );
   assert.doesNotMatch(
-    `${loginPageSource}\n${loginControllerSource}`,
+    `${loginPageSource}\n${loginControllerSource}\n${loginOAuthCodeExchangeSource}`,
     /window\.location\.replace\(buildAuthFinishUrl/,
     'login should not route through an extra auth finish page'
   );
   assert.doesNotMatch(
-    `${loginPageSource}\n${loginControllerSource}`,
+    `${loginPageSource}\n${loginControllerSource}\n${loginOAuthCodeExchangeSource}`,
     /\/auth\/finish/,
     'login should not add another intermediate auth page'
   );
@@ -46,12 +47,12 @@ test('OAuth cookie fallback redirects only after a new Supabase auth cookie appe
 
 test('login waits for the final next path before starting PKCE exchange', () => {
   assert.match(
-    loginControllerSource,
+    loginOAuthCodeExchangeSource,
     /if \(!nextPathReady\) return;\s+if \(typeof window === 'undefined'\) return;\s+const params = new URLSearchParams\(window\.location\.search\);/,
     'PKCE exchange should not start until nextPath has been resolved'
   );
   assert.match(
-    loginControllerSource,
+    loginOAuthCodeExchangeSource,
     /oauthCodeExchangeStartedRef\.current = true;[\s\S]*const target = sanitizeNextPath\(params\.get\('next'\) \?\? nextPath\);/,
     'the exchange should capture a stable redirect target after nextPathReady'
   );
