+++ tests/login-page-architecture.test.ts
@@ -10,6 +10,9 @@ const autofillHookPath = join(root, 'frontend/app/(core)/login/_hooks/useLoginAu
 const browserLocaleHookPath = join(root, 'frontend/app/(core)/login/_hooks/useLoginBrowserLocale.ts');
 const modeQueryHookPath = join(root, 'frontend/app/(core)/login/_hooks/useLoginModeFromQuery.ts');
 const nextTargetHookPath = join(root, 'frontend/app/(core)/login/_hooks/useLoginNextTarget.ts');
+const authenticatedRedirectHookPath = join(root, 'frontend/app/(core)/login/_hooks/useLoginAuthenticatedRedirect.ts');
+const authHashSessionHookPath = join(root, 'frontend/app/(core)/login/_hooks/useLoginAuthHashSession.ts');
+const oauthCodeExchangeHookPath = join(root, 'frontend/app/(core)/login/_hooks/useLoginOAuthCodeExchange.ts');
 const copyPath = join(root, 'frontend/app/(core)/login/_lib/login-copy.ts');
 const helpersPath = join(root, 'frontend/app/(core)/login/_lib/login-helpers.ts');
 const authSurfacePath = join(root, 'frontend/app/(core)/login/_components/LoginAuthSurface.tsx');
@@ -18,6 +21,9 @@ const pageSource = readFileSync(pagePath, 'utf8');
 const controllerSource = readFileSync(controllerPath, 'utf8');
 const autofillHookSource = readFileSync(autofillHookPath, 'utf8');
 const nextTargetHookSource = readFileSync(nextTargetHookPath, 'utf8');
+const authenticatedRedirectHookSource = readFileSync(authenticatedRedirectHookPath, 'utf8');
+const authHashSessionHookSource = readFileSync(authHashSessionHookPath, 'utf8');
+const oauthCodeExchangeHookSource = readFileSync(oauthCodeExchangeHookPath, 'utf8');
 
 test('login page delegates localized copy and browser helpers to route-local modules', () => {
   assert.ok(existsSync(controllerPath), 'login controller hook should stay in a route-local hook module');
@@ -25,6 +31,9 @@ test('login page delegates localized copy and browser helpers to route-local mod
   assert.ok(existsSync(browserLocaleHookPath), 'login browser locale should stay in a route-local hook module');
   assert.ok(existsSync(modeQueryHookPath), 'login mode query sync should stay in a route-local hook module');
   assert.ok(existsSync(nextTargetHookPath), 'login next target state should stay in a route-local hook module');
+  assert.ok(existsSync(authenticatedRedirectHookPath), 'login authenticated redirect probes should stay in a route-local hook module');
+  assert.ok(existsSync(authHashSessionHookPath), 'login hash session handling should stay in a route-local hook module');
+  assert.ok(existsSync(oauthCodeExchangeHookPath), 'login OAuth code exchange should stay in a route-local hook module');
   assert.ok(existsSync(copyPath), 'login localized copy should stay in a route-local copy module');
   assert.ok(existsSync(helpersPath), 'login browser helpers should stay in a route-local helper module');
   assert.ok(existsSync(authSurfacePath), 'login form UI should stay in a route-local component module');
@@ -45,6 +54,9 @@ test('login page delegates localized copy and browser helpers to route-local mod
   assert.match(controllerSource, /from '\.\/useLoginBrowserLocale'/);
   assert.match(controllerSource, /from '\.\/useLoginModeFromQuery'/);
   assert.match(controllerSource, /from '\.\/useLoginNextTarget'/);
+  assert.match(controllerSource, /from '\.\/useLoginAuthenticatedRedirect'/);
+  assert.match(controllerSource, /from '\.\/useLoginAuthHashSession'/);
+  assert.match(controllerSource, /from '\.\/useLoginOAuthCodeExchange'/);
 });
 
 test('login page does not regain auth copy or browser helper ownership', () => {
@@ -79,8 +91,14 @@ test('login helper modules expose the expected route contract', () => {
   assert.match(nextTargetHookSource, /export function useLoginNextTarget/, 'next target resolution should live in its own hook');
   assert.match(nextTargetHookSource, /LOGIN_NEXT_STORAGE_KEY/, 'next target hook should own login target storage');
   assert.match(nextTargetHookSource, /LOGIN_LAST_TARGET_KEY/, 'next target hook should own last target storage');
-  assert.match(controllerSource, /exchangeCodeForSession\(oauthCode\)/, 'controller hook should own browser-side PKCE exchange');
-  assert.match(controllerSource, /startOAuthCookieRedirectFallback/, 'controller hook should own OAuth cookie fallback wiring');
+  assert.doesNotMatch(controllerSource, /exchangeCodeForSession\(oauthCode\)/, 'OAuth PKCE exchange belongs in useLoginOAuthCodeExchange');
+  assert.doesNotMatch(controllerSource, /startOAuthCookieRedirectFallback/, 'OAuth cookie fallback wiring belongs in useLoginOAuthCodeExchange');
+  assert.match(oauthCodeExchangeHookSource, /exchangeCodeForSession\(oauthCode\)/, 'OAuth code exchange hook should own browser-side PKCE exchange');
+  assert.match(oauthCodeExchangeHookSource, /startOAuthCookieRedirectFallback/, 'OAuth code exchange hook should own OAuth cookie fallback wiring');
+  assert.match(authenticatedRedirectHookSource, /export function useLoginAuthenticatedRedirect/, 'authenticated redirect hook should be exported');
+  assert.match(authenticatedRedirectHookSource, /async function redirectFromExistingBrowserSession\(target: string\): Promise<boolean>/, 'authenticated redirect hook should expose the Safari fallback redirect');
+  assert.match(authHashSessionHookSource, /export function useLoginAuthHashSession/, 'auth hash session hook should be exported');
+  assert.match(authHashSessionHookSource, /setSession\(\{/, 'auth hash session hook should own Supabase hash session hydration');
   assert.match(controllerSource, /signInWithOAuth/, 'controller hook should own Google OAuth submission');
   assert.match(copySource, /export const AUTH_COPY =/, 'copy module should export the auth copy dictionary');
   assert.match(copySource, /export type AuthMode =/, 'copy module should export the auth mode type');
